pyproject.toml
[tool.poetry]
name = "fastapi_realworld"
version = "0.1.0"
description = "FastAPI Real World description"
authors = ["Medniy200 Team <your@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.10"
fastapi = "^0.70.0"
environs = "^9.3.5"
uvicorn = "0.16.0"
alembic = "1.7.5"
asyncpg = "^0.25.0"
psycopg2 = "^2.9.2"
gino = {extras = ["pg", "starlette"], version = "^1.0.1"}
loguru = "^0.5.3"
python-slugify = "^5.0.2"
furl = "^2.1.3"
slugify = "^0.0.1"
Authlib = "^0.15.5"
itsdangerous = "^2.1.0"
httpx = "^0.22.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uuid = "^1.30"


[tool.poetry.dev-dependencies]
pytest = "^6.1.0"
pytest-asyncio = "^0.18.3"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
mypy = "^0.931"
Sphinx = "^4.3.2"
black = {version = "^22.8.0)", python = ">=3.9.10"}
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "-p no:warnings --asyncio-mode=strict"
testpaths = ["tests"]


[tool.black]
line-length = 110
target-version = ['py39']



